Detect and correct input file format errors
The question pool files have lots of formatting errors in them. Some we already
catch and fix automatically. For the rest, we now detect and print them while
parsing the file.

Rather than update the script to fix every conceivable error, we'll just edit
the input files to fix the handful that appear.